feat: `util/are_part_min_max_increasing.py` tool
Add a tool to check if row groups .min / .max for a particular column (eg url_surtkey) are increasing within each parquet file passed as input.

Note that this tool only checks that, within a single `.parquet` file, each row group's `.min` is greater than or equal to the previous row group's `.max`; further, the context of this condition is restricted to a single parquet file. For cases where the table as a whole consists of multiple parquet files, the condition may not hold across file boundaries.

Signed-off-by: Damian Stewart <ot@damianstewart.com>
Co-authored-by: Damian Stewart <ot@damianstewart.com>